% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{create_campaign}
\alias{create_campaign}
\title{Creates or updates a campaign}
\usage{
create_campaign(ApplicationId, WriteCampaignRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{WriteCampaignRequest}{}
}
\description{
Creates or updates a campaign.
}
\examples{

}
