% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{create_segment}
\alias{create_segment}
\title{Used to create or update a segment}
\usage{
create_segment(ApplicationId, WriteSegmentRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{WriteSegmentRequest}{}
}
\description{
Used to create or update a segment.
}
\examples{

}
