% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_adm_channel}
\alias{delete_adm_channel}
\title{Delete an ADM channel}
\usage{
delete_adm_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}
}
\description{
Delete an ADM channel.
}
\examples{

}
