% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_apns_channel}
\alias{delete_apns_channel}
\title{Deletes the APNs channel for an app}
\usage{
delete_apns_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Deletes the APNs channel for an app.
}
\section{Accepted Parameters}{
\preformatted{delete_apns_channel(
  ApplicationId = "string"
)
}
}

