% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_apns_voip_channel}
\alias{delete_apns_voip_channel}
\title{Delete an APNS VoIP channel}
\usage{
delete_apns_voip_channel(
  ApplicationId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Delete an APNS VoIP channel
}
\examples{

}
