% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_apns_voip_sandbox_channel}
\alias{delete_apns_voip_sandbox_channel}
\title{Delete an APNS VoIP sandbox channel}
\usage{
delete_apns_voip_sandbox_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Delete an APNS VoIP sandbox channel
}
\section{Accepted Parameters}{
\preformatted{delete_apns_voip_sandbox_channel(
  ApplicationId = "string"
)
}
}

