% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_app}
\alias{delete_app}
\title{Deletes an app}
\usage{
delete_app(
  ApplicationId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Deletes an app.
}
\examples{

}
