% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_baidu_channel}
\alias{delete_baidu_channel}
\title{Delete a BAIDU GCM channel}
\usage{
delete_baidu_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Delete a BAIDU GCM channel
}
\section{Accepted Parameters}{
\preformatted{delete_baidu_channel(
  ApplicationId = "string"
)
}
}

