% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_campaign}
\alias{delete_campaign}
\title{Deletes a campaign}
\usage{
delete_campaign(
  ApplicationId = "string",
  CampaignId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{CampaignId}{[required] The unique ID of the campaign.}
}
\description{
Deletes a campaign.
}
\examples{

}
