% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_email_channel}
\alias{delete_email_channel}
\title{Delete an email channel}
\usage{
delete_email_channel(
  ApplicationId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Delete an email channel.
}
\examples{

}
