% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_event_stream}
\alias{delete_event_stream}
\title{Deletes the event stream for an app}
\usage{
delete_event_stream(ApplicationId)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}
}
\description{
Deletes the event stream for an app.
}
\examples{

}
