% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_gcm_channel}
\alias{delete_gcm_channel}
\title{Deletes the GCM channel for an app}
\usage{
delete_gcm_channel(
  ApplicationId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Deletes the GCM channel for an app.
}
\examples{

}
