% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_segment}
\alias{delete_segment}
\title{Deletes a segment}
\usage{
delete_segment(SegmentId, ApplicationId)
}
\arguments{
\item{SegmentId}{The unique ID of the segment.}

\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}
}
\description{
Deletes a segment.
}
\examples{

}
