% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_sms_channel}
\alias{delete_sms_channel}
\title{Delete an SMS channel}
\usage{
delete_sms_channel(
  ApplicationId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Delete an SMS channel.
}
\examples{

}
