% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_user_endpoints}
\alias{delete_user_endpoints}
\title{Deletes endpoints that are associated with a User ID}
\usage{
delete_user_endpoints(ApplicationId, UserId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{UserId}{[required] The unique ID of the user.}
}
\description{
Deletes endpoints that are associated with a User ID.
}
\section{Accepted Parameters}{
\preformatted{delete_user_endpoints(
  ApplicationId = "string",
  UserId = "string"
)
}
}

