% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_voice_channel}
\alias{delete_voice_channel}
\title{Delete an Voice channel}
\usage{
delete_voice_channel(
  ApplicationId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Delete an Voice channel
}
\examples{

}
