% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_adm_channel}
\alias{get_adm_channel}
\title{Get an ADM channel}
\usage{
get_adm_channel(
  ApplicationId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get an ADM channel.
}
\examples{

}
