% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_apns_channel}
\alias{get_apns_channel}
\title{Returns information about the APNs channel for an app}
\usage{
get_apns_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}
}
\description{
Returns information about the APNs channel for an app.
}
\examples{

}
