% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_apns_voip_channel}
\alias{get_apns_voip_channel}
\title{Get an APNS VoIP channel}
\usage{
get_apns_voip_channel(
  ApplicationId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get an APNS VoIP channel
}
\examples{

}
