% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_apns_voip_sandbox_channel}
\alias{get_apns_voip_sandbox_channel}
\title{Get an APNS VoIPSandbox channel}
\usage{
get_apns_voip_sandbox_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get an APNS VoIPSandbox channel
}
\section{Accepted Parameters}{
\preformatted{get_apns_voip_sandbox_channel(
  ApplicationId = "string"
)
}
}

