% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_application_settings}
\alias{get_application_settings}
\title{Used to request the settings for an app}
\usage{
get_application_settings(
  ApplicationId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Used to request the settings for an app.
}
\examples{

}
