% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_apps}
\alias{get_apps}
\title{Returns information about your apps}
\usage{
get_apps(PageSize = NULL, Token = NULL)
}
\arguments{
\item{PageSize}{The number of entries you want on each page in the response.}

\item{Token}{The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns information about your apps.
}
\examples{

}
