% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_baidu_channel}
\alias{get_baidu_channel}
\title{Get a BAIDU GCM channel}
\usage{
get_baidu_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get a BAIDU GCM channel
}
\section{Accepted Parameters}{
\preformatted{get_baidu_channel(
  ApplicationId = "string"
)
}
}

