% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_channels}
\alias{get_channels}
\title{Get all channels}
\usage{
get_channels(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get all channels.
}
\section{Accepted Parameters}{
\preformatted{get_channels(
  ApplicationId = "string"
)
}
}

