% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_email_channel}
\alias{get_email_channel}
\title{Get an email channel}
\usage{
get_email_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get an email channel.
}
\section{Accepted Parameters}{
\preformatted{get_email_channel(
  ApplicationId = "string"
)
}
}

