% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_event_stream}
\alias{get_event_stream}
\title{Returns the event stream for an app}
\usage{
get_event_stream(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Returns the event stream for an app.
}
\section{Accepted Parameters}{
\preformatted{get_event_stream(
  ApplicationId = "string"
)
}
}

