% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_export_job}
\alias{get_export_job}
\title{Returns information about an export job}
\usage{
get_export_job(ApplicationId, JobId)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{JobId}{The unique ID of the job.}
}
\description{
Returns information about an export job.
}
\examples{

}
