% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_export_jobs}
\alias{get_export_jobs}
\title{Returns information about your export jobs}
\usage{
get_export_jobs(ApplicationId, PageSize = NULL, Token = NULL)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{PageSize}{The number of entries you want on each page in the response.}

\item{Token}{The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns information about your export jobs.
}
\examples{

}
