% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_gcm_channel}
\alias{get_gcm_channel}
\title{Returns information about the GCM channel for an app}
\usage{
get_gcm_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Returns information about the GCM channel for an app.
}
\section{Accepted Parameters}{
\preformatted{get_gcm_channel(
  ApplicationId = "string"
)
}
}

