% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_import_job}
\alias{get_import_job}
\title{Returns information about an import job}
\usage{
get_import_job(ApplicationId, JobId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{JobId}{[required] The unique ID of the job.}
}
\description{
Returns information about an import job.
}
\section{Accepted Parameters}{
\preformatted{get_import_job(
  ApplicationId = "string",
  JobId = "string"
)
}
}

