% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_segment}
\alias{get_segment}
\title{Returns information about a segment}
\usage{
get_segment(SegmentId, ApplicationId)
}
\arguments{
\item{SegmentId}{The unique ID of the segment.}

\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}
}
\description{
Returns information about a segment.
}
\examples{

}
