% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_segment_import_jobs}
\alias{get_segment_import_jobs}
\title{Returns a list of import jobs for a specific segment}
\usage{
get_segment_import_jobs(ApplicationId, PageSize = NULL, SegmentId,
  Token = NULL)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{PageSize}{The number of entries you want on each page in the response.}

\item{SegmentId}{[required] The unique ID of the segment.}

\item{Token}{The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns a list of import jobs for a specific segment.
}
\section{Accepted Parameters}{
\preformatted{get_segment_import_jobs(
  ApplicationId = "string",
  PageSize = "string",
  SegmentId = "string",
  Token = "string"
)
}
}

