% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_segment_version}
\alias{get_segment_version}
\title{Returns information about a segment version}
\usage{
get_segment_version(SegmentId, Version, ApplicationId)
}
\arguments{
\item{SegmentId}{The unique ID of the segment.}

\item{Version}{The segment version.}

\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}
}
\description{
Returns information about a segment version.
}
\examples{

}
