% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_segment_versions}
\alias{get_segment_versions}
\title{Returns information about your segment versions}
\usage{
get_segment_versions(SegmentId, ApplicationId, PageSize = NULL,
  Token = NULL)
}
\arguments{
\item{SegmentId}{The unique ID of the segment.}

\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{PageSize}{The number of entries you want on each page in the response.}

\item{Token}{The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns information about your segment versions.
}
\examples{

}
