% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_segments}
\alias{get_segments}
\title{Used to get information about your segments}
\usage{
get_segments(ApplicationId, PageSize = NULL, Token = NULL)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{PageSize}{The number of entries you want on each page in the response.}

\item{Token}{The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Used to get information about your segments.
}
\section{Accepted Parameters}{
\preformatted{get_segments(
  ApplicationId = "string",
  PageSize = "string",
  Token = "string"
)
}
}

