% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_package.R
\docType{package}
\name{paws.pinpoint-package}
\alias{paws.pinpoint}
\alias{paws.pinpoint-package}
\title{paws.pinpoint: Amazon Pinpoint}
\description{
Amazon Pinpoint
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
