% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{put_events}
\alias{put_events}
\title{Use to record events for endpoints}
\usage{
put_events(ApplicationId, EventsRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{EventsRequest}{}
}
\description{
Use to record events for endpoints. This method creates events and creates or updates the endpoints that those events are associated with.
}
\examples{

}
