% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{remove_attributes}
\alias{remove_attributes}
\title{Used to remove the attributes for an app}
\usage{
remove_attributes(AttributeType, ApplicationId, UpdateAttributesRequest)
}
\arguments{
\item{AttributeType}{Type of attribute. Can be endpoint-custom-attributes, endpoint-custom-metrics, endpoint-user-attributes.}

\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{UpdateAttributesRequest}{}
}
\description{
Used to remove the attributes for an app
}
\examples{

}
