% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{send_messages}
\alias{send_messages}
\title{Used to send a direct message}
\usage{
send_messages(ApplicationId, MessageRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{MessageRequest}{}
}
\description{
Used to send a direct message.
}
\examples{

}
