% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{send_users_messages}
\alias{send_users_messages}
\title{Used to send a message to a list of users}
\usage{
send_users_messages(ApplicationId, SendUsersMessageRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{SendUsersMessageRequest}{}
}
\description{
Used to send a message to a list of users.
}
\examples{

}
