% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_adm_channel}
\alias{update_adm_channel}
\title{Update an ADM channel}
\usage{
update_adm_channel(ApplicationId, ADMChannelRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{ADMChannelRequest}{}
}
\description{
Update an ADM channel.
}
\examples{

}
