% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_apns_channel}
\alias{update_apns_channel}
\title{Use to update the APNs channel for an app}
\usage{
update_apns_channel(ApplicationId, APNSChannelRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{APNSChannelRequest}{}
}
\description{
Use to update the APNs channel for an app.
}
\examples{

}
