% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_apns_voip_channel}
\alias{update_apns_voip_channel}
\title{Update an APNS VoIP channel}
\usage{
update_apns_voip_channel(ApplicationId, APNSVoipChannelRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{APNSVoipChannelRequest}{}
}
\description{
Update an APNS VoIP channel
}
\examples{

}
