% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_apns_voip_sandbox_channel}
\alias{update_apns_voip_sandbox_channel}
\title{Update an APNS VoIP sandbox channel}
\usage{
update_apns_voip_sandbox_channel(ApplicationId, APNSVoipSandboxChannelRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{APNSVoipSandboxChannelRequest}{}
}
\description{
Update an APNS VoIP sandbox channel
}
\examples{

}
