% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_application_settings}
\alias{update_application_settings}
\title{Used to update the settings for an app}
\usage{
update_application_settings(ApplicationId, WriteApplicationSettingsRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{WriteApplicationSettingsRequest}{}
}
\description{
Used to update the settings for an app.
}
\examples{

}
