% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_baidu_channel}
\alias{update_baidu_channel}
\title{Update a BAIDU GCM channel}
\usage{
update_baidu_channel(ApplicationId, BaiduChannelRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{BaiduChannelRequest}{}
}
\description{
Update a BAIDU GCM channel
}
\examples{

}
