% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_campaign}
\alias{update_campaign}
\title{Use to update a campaign}
\usage{
update_campaign(CampaignId, ApplicationId, WriteCampaignRequest)
}
\arguments{
\item{CampaignId}{The unique ID of the campaign.}

\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{WriteCampaignRequest}{}
}
\description{
Use to update a campaign.
}
\examples{

}
