% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_email_channel}
\alias{update_email_channel}
\title{Update an email channel}
\usage{
update_email_channel(ApplicationId, EmailChannelRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{EmailChannelRequest}{[required]}
}
\description{
Update an email channel.
}
\section{Accepted Parameters}{
\preformatted{update_email_channel(
  ApplicationId = "string",
  EmailChannelRequest = list(
    ConfigurationSet = "string",
    Enabled = TRUE|FALSE,
    FromAddress = "string",
    Identity = "string",
    RoleArn = "string"
  )
)
}
}

