% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_endpoint}
\alias{update_endpoint}
\title{Creates or updates an endpoint}
\usage{
update_endpoint(ApplicationId, EndpointId, EndpointRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{EndpointId}{The unique ID of the endpoint.}

\item{EndpointRequest}{}
}
\description{
Creates or updates an endpoint.
}
\examples{

}
