% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_endpoints_batch}
\alias{update_endpoints_batch}
\title{Use to update a batch of endpoints}
\usage{
update_endpoints_batch(ApplicationId, EndpointBatchRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{EndpointBatchRequest}{}
}
\description{
Use to update a batch of endpoints.
}
\examples{

}
