% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_gcm_channel}
\alias{update_gcm_channel}
\title{Use to update the GCM channel for an app}
\usage{
update_gcm_channel(ApplicationId, GCMChannelRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{GCMChannelRequest}{}
}
\description{
Use to update the GCM channel for an app.
}
\examples{

}
