% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_segment}
\alias{update_segment}
\title{Used to update a segment}
\usage{
update_segment(SegmentId, ApplicationId, WriteSegmentRequest)
}
\arguments{
\item{SegmentId}{The unique ID of the segment.}

\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{WriteSegmentRequest}{}
}
\description{
Used to update a segment.
}
\examples{

}
