% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{create_app}
\alias{create_app}
\title{Creates or updates an app}
\usage{
create_app(
  CreateApplicationRequest = list(
    Name = "string"
  )
)
}
\arguments{
\item{CreateApplicationRequest}{[required]}
}
\description{
Creates or updates an app.
}
\examples{

}
