% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{create_export_job}
\alias{create_export_job}
\title{Creates an export job}
\usage{
create_export_job(ApplicationId, ExportJobRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{ExportJobRequest}{}
}
\description{
Creates an export job.
}
\examples{

}
