% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{create_import_job}
\alias{create_import_job}
\title{Creates or updates an import job}
\usage{
create_import_job(ApplicationId, ImportJobRequest)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{ImportJobRequest}{}
}
\description{
Creates or updates an import job.
}
\examples{

}
