% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_campaign}
\alias{delete_campaign}
\title{Deletes a campaign}
\usage{
delete_campaign(CampaignId, ApplicationId)
}
\arguments{
\item{CampaignId}{The unique ID of the campaign.}

\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}
}
\description{
Deletes a campaign.
}
\examples{

}
