% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_endpoint}
\alias{delete_endpoint}
\title{Deletes an endpoint}
\usage{
delete_endpoint(ApplicationId, EndpointId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{EndpointId}{[required] The unique ID of the endpoint.}
}
\description{
Deletes an endpoint.
}
\section{Accepted Parameters}{
\preformatted{delete_endpoint(
  ApplicationId = "string",
  EndpointId = "string"
)
}
}

