% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{delete_segment}
\alias{delete_segment}
\title{Deletes a segment}
\usage{
delete_segment(
  ApplicationId = "string",
  SegmentId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{SegmentId}{[required] The unique ID of the segment.}
}
\description{
Deletes a segment.
}
\examples{

}
