% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_campaign}
\alias{get_campaign}
\title{Returns information about a campaign}
\usage{
get_campaign(CampaignId, ApplicationId)
}
\arguments{
\item{CampaignId}{The unique ID of the campaign.}

\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}
}
\description{
Returns information about a campaign.
}
\examples{

}
