% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_campaign}
\alias{get_campaign}
\title{Returns information about a campaign}
\usage{
get_campaign(ApplicationId, CampaignId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{CampaignId}{[required] The unique ID of the campaign.}
}
\description{
Returns information about a campaign.
}
\section{Accepted Parameters}{
\preformatted{get_campaign(
  ApplicationId = "string",
  CampaignId = "string"
)
}
}

