% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_campaign_version}
\alias{get_campaign_version}
\title{Returns information about a specific version of a campaign}
\usage{
get_campaign_version(Version, ApplicationId, CampaignId)
}
\arguments{
\item{Version}{The version of the campaign.}

\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{CampaignId}{The unique ID of the campaign.}
}
\description{
Returns information about a specific version of a campaign.
}
\examples{

}
