% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Returns information about an endpoint}
\usage{
get_endpoint(ApplicationId, EndpointId)
}
\arguments{
\item{ApplicationId}{The unique ID of your Amazon Pinpoint application.}

\item{EndpointId}{The unique ID of the endpoint.}
}
\description{
Returns information about an endpoint.
}
\examples{

}
