% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_segment}
\alias{get_segment}
\title{Returns information about a segment}
\usage{
get_segment(ApplicationId, SegmentId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{SegmentId}{[required] The unique ID of the segment.}
}
\description{
Returns information about a segment.
}
\section{Accepted Parameters}{
\preformatted{get_segment(
  ApplicationId = "string",
  SegmentId = "string"
)
}
}

