% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_sms_channel}
\alias{get_sms_channel}
\title{Get an SMS channel}
\usage{
get_sms_channel(
  ApplicationId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get an SMS channel.
}
\examples{

}
