% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{get_voice_channel}
\alias{get_voice_channel}
\title{Get a Voice Channel}
\usage{
get_voice_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get a Voice Channel
}
\section{Accepted Parameters}{
\preformatted{get_voice_channel(
  ApplicationId = "string"
)
}
}

