% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{remove_attributes}
\alias{remove_attributes}
\title{Used to remove the attributes for an app}
\usage{
remove_attributes(
  ApplicationId = "string",
  AttributeType = "string",
  UpdateAttributesRequest = list(
    Blacklist = list(
      "string"
    )
  )
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{AttributeType}{[required] Type of attribute. Can be endpoint-custom-attributes, endpoint-custom-metrics, endpoint-user-attributes.}

\item{UpdateAttributesRequest}{[required]}
}
\description{
Used to remove the attributes for an app
}
\examples{

}
