% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_adm_channel}
\alias{update_adm_channel}
\title{Update an ADM channel}
\usage{
update_adm_channel(
  ADMChannelRequest = list(
    ClientId = "string",
    ClientSecret = "string",
    Enabled = TRUE|FALSE
  ),
  ApplicationId = "string"
)
}
\arguments{
\item{ADMChannelRequest}{[required]}

\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Update an ADM channel.
}
\examples{

}
