% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_apns_sandbox_channel}
\alias{update_apns_sandbox_channel}
\title{Update an APNS sandbox channel}
\usage{
update_apns_sandbox_channel(
  APNSSandboxChannelRequest = list(
    BundleId = "string",
    Certificate = "string",
    DefaultAuthenticationMethod = "string",
    Enabled = TRUE|FALSE,
    PrivateKey = "string",
    TeamId = "string",
    TokenKey = "string",
    TokenKeyId = "string"
  ),
  ApplicationId = "string"
)
}
\arguments{
\item{APNSSandboxChannelRequest}{[required]}

\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Update an APNS sandbox channel.
}
\examples{

}
