% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpoint_operations.R
\name{update_voice_channel}
\alias{update_voice_channel}
\title{Update an Voice channel}
\usage{
update_voice_channel(
  ApplicationId = "string",
  VoiceChannelRequest = list(
    Enabled = TRUE|FALSE
  )
)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{VoiceChannelRequest}{[required]}
}
\description{
Update an Voice channel
}
\examples{

}
